+++ common/src/main/java/com/gitlab/cdagaming/craftpresence/config/gui/PresenceEditorGui.java
@@ -569,7 +569,7 @@ public class PresenceEditorGui extends ConfigurationGui<PresenceData> {
     }
 
     @Override
-    protected void mouseClicked(int mouseX, int mouseY, int mouseButton) {
+    protected void mouseClicked(float mouseX, float mouseY, int mouseButton) {
         super.mouseClicked(mouseX, mouseY, mouseButton);
         visualizer.onClick(childFrame, mouseX, mouseY, mouseButton);
     }
